+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Enh: Use default timezone instead of deprecated server timezone
 - Enh: Added French translations
 - Fix #47: Error adding an iCalendar link (updated ics-parser library to version 3)
+- Enh: Validation added for better understanding
 
 1.3.3 (March 7, 2023)
 ---------------------